Refactor: change tab buuton location from page to common
